refactored the code to Daniel's suggestions
